+++ CHANGELOG.md
@@ -10,6 +10,7 @@ For details on the changes in each release, see [the Releases page](https://gith
 - the `mail_overrides` directory has been renamed to `mail`
 - the `templates_overrides` directory has been renamed to `templates`
 - the `overrides` directory has been renamed to `domain_overrides`
+- the `[smtp]ssl_verify` option has changed from `"true"/"false"` to `true/false`
 
 ### 1.6 -> 1.7